fix(activation): fix type error in forward() method
Edit isinstance (..., int) with isinstance (..., float) in forward()
method in ReLu class. Because output parameter comes as float.

+++ activation/relu.py
@@ -6,7 +6,7 @@ class ReLu():
         return
     
     def forward(self, output):
-        if isinstance(output, int):
+        if isinstance(output, float):
             return np.max(0, output)
         elif isinstance(output, np.ndarray):
             return np.maximum(np.zeros(output.size), output)
